requirements.txt
requests
certifi
confluent-kafka[avro,json,protobuf]
websocket-client
rel
pyspark
